add authors to biapy card
Hey there! 

I just added some authors to the biapy app for the zoo as I think it would be nicer to see all of you in the [main card](https://bioimage.io/#/?partner=biapy&type=application&id=biapy%2Fbiapy).

Otherwise, we are not really listing you anywhere (this should also be changed).